requirements.txt
telethon
mysql-connector-python